perf(commonpb): inline Timestamp as fixed64 scalar with typed accessors
Every proto message that carries a timestamp previously stored
`*commonpb.Timestamp` — a pointer to a heap-allocated single-uint64 message.
On the hot path each unmarshal allocated one struct per timestamp field; a
typical Transaction carries 4 timestamps, so a single decoded log entry
generated ≥4 avoidable heap allocations before any processing began.

Timestamp is now a `fixed64` scalar directly on each parent message, wrapped
Go-side by a value type `type commonpb.Timestamp uint64` carrying the domain
methods (AsTime, IsSet, MarshalJSON, Micros). The zero value doubles as the
"unset" sentinel for nullable fields (RevertedAt, DeletedAt); the HLC never
emits micros=0 in practice so this collapse is safe.

Ergonomic accessors:
- misc/proto/ledger_options.proto declares a custom field option
  `[(ledger.value_type) = "commonpb.Timestamp"]`.
- protoc-gen-reader now emits `<Field>Ts() commonpb.Timestamp` on messages
  whose scalar fields carry this annotation, so callers write
  `tx.TimestampTs().AsTime()` instead of `commonpb.Timestamp(tx.GetTimestamp()).AsTime()`.
- `Ts` suffix avoids the Go field/method naming conflict on the exported
  `Timestamp uint64` proto field.

Other bits:
- `Transaction.UnmarshalJSON` added, mirroring the hand-written MarshalJSON:
  RFC3339 strings ↔ fixed64 micros — otherwise round-trip through
  HydrateLog would fail on CreatedTransaction / RevertedTransaction payloads.
- `NewTimestamp(t <= epoch)` collapses to TimestampUnset (0) rather than
  panicking; the constructor is now safe on untrusted input paths
  (LedgerLog.UnmarshalJSON, JSON API).

Wire format changes: `common.Timestamp foo = N;` (wire type 2, length-
delimited) → `fixed64 foo = N;` (wire type 1). Not backward-compatible; the
audit chain hash over marshaled proto differs for the same logical entry.
v3-only cutover, no migration path for chains persisted before this change.

+++ internal/application/admission/admission_test.go
@@ -48,7 +48,7 @@ func createTestStore(t *testing.T) *dal.Store {
 	info := &commonpb.LedgerInfo{
 		Name:      testLedgerName,
 		Id:        1,
-		CreatedAt: commonpb.NewTimestamp(time.Now()),
+		CreatedAt: uint64(commonpb.NewTimestamp(time.Now())),
 	}
 	batch := s.OpenWriteSession()
 	require.NoError(t, state.SaveLedger(batch, info))
