+++ internal/application/events/sink_nats_integration_test.go
@@ -139,7 +139,7 @@ func TestNATSSinkIntegration_PublishAndConsume(t *testing.T) {
 			Payload: &commonpb.LogPayload{
 				Type: &commonpb.LogPayload_CreateLedger{
 					CreateLedger: &commonpb.CreatedLedgerLog{
-						Name: "orders", CreatedAt: commonpb.NewTimestamp(now),
+						Name: "orders", CreatedAt: uint64(commonpb.NewTimestamp(now)),
 					},
 				},
 			},
@@ -343,7 +343,7 @@ func TestNATSSinkIntegration_SubjectRouting(t *testing.T) {
 			Payload: &commonpb.LogPayload{
 				Type: &commonpb.LogPayload_CreateLedger{
 					CreateLedger: &commonpb.CreatedLedgerLog{
-						Name: "orders", CreatedAt: commonpb.NewTimestamp(now),
+						Name: "orders", CreatedAt: uint64(commonpb.NewTimestamp(now)),
 					},
 				},
 			},
@@ -354,7 +354,7 @@ func TestNATSSinkIntegration_SubjectRouting(t *testing.T) {
 			Payload: &commonpb.LogPayload{
 				Type: &commonpb.LogPayload_CreateLedger{
 					CreateLedger: &commonpb.CreatedLedgerLog{
-						Name: "payments", CreatedAt: commonpb.NewTimestamp(now),
+						Name: "payments", CreatedAt: uint64(commonpb.NewTimestamp(now)),
 					},
 				},
 			},
