+++ internal/infra/state/archiver_test.go
@@ -217,7 +217,7 @@ func TestArchiverArchivesAndProposes(t *testing.T) {
 	batch := dataStore.OpenWriteSession()
 	require.NoError(t, AppendLogs(batch, []*commonpb.Log{{
 		Sequence: 1,
-		Payload:  &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{CreateLedger: &commonpb.CreatedLedgerLog{Name: "test", CreatedAt: commonpb.NewTimestamp(libtime.Now())}}},
+		Payload:  &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{CreateLedger: &commonpb.CreatedLedgerLog{Name: "test", CreatedAt: uint64(commonpb.NewTimestamp(libtime.Now()))}}},
 	}}))
 	require.NoError(t, batch.Commit())
 
@@ -460,7 +460,7 @@ func TestArchiverNonLeaderRetries(t *testing.T) {
 	batch := dataStore.OpenWriteSession()
 	require.NoError(t, AppendLogs(batch, []*commonpb.Log{{
 		Sequence: 1,
-		Payload:  &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{CreateLedger: &commonpb.CreatedLedgerLog{Name: "test", CreatedAt: commonpb.NewTimestamp(libtime.Now())}}},
+		Payload:  &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{CreateLedger: &commonpb.CreatedLedgerLog{Name: "test", CreatedAt: uint64(commonpb.NewTimestamp(libtime.Now()))}}},
 	}}))
 	require.NoError(t, batch.Commit())
 
@@ -488,7 +488,7 @@ func TestArchiverSSTRoundtrip(t *testing.T) {
 		Sequence: 1,
 		Payload: &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{
 			CreateLedger: &commonpb.CreatedLedgerLog{
-				Name: "test-ledger", CreatedAt: commonpb.NewTimestamp(libtime.Now()),
+				Name: "test-ledger", CreatedAt: uint64(commonpb.NewTimestamp(libtime.Now())),
 			},
 		}},
 	}}))
@@ -566,7 +566,7 @@ func archiveRequestForTest(t *testing.T) (*dal.Store, ArchiveRequest) {
 		Payload: &commonpb.LogPayload{Type: &commonpb.LogPayload_CreateLedger{
 			CreateLedger: &commonpb.CreatedLedgerLog{
 				Name:      "test-ledger",
-				CreatedAt: commonpb.NewTimestamp(libtime.Now()),
+				CreatedAt: uint64(commonpb.NewTimestamp(libtime.Now())),
 			},
 		}},
 	}}))
